Make TranslationPrompts a class and TranslatablePrompts a module because it seemed that we weren’t able to use it as both. (i.e. couldn’t call the module code from the rake task)
