+++ services/QuillLMS/lib/tasks/temporary/translate.rake
@@ -45,7 +45,7 @@ namespace :translate do
     result = OpenAI::Translate.run(english_text:, prompt:)
     filename = "#{fileroot}.#{locale}.#{fileformat}"
     puts result
-    File.write(Rails.root.join(filefolder, filename), result)
+    File.write(Rails.root.join(filefolder, 'translations', filename), result)
   end
 
   def fileroot = 'feedback_strings'
@@ -53,7 +53,7 @@ namespace :translate do
   def filefolder = 'config/locales'
 
   def feedback_data
-    file_path = Rails.root.join(filefolder, 'translations', "#{fileroot}.#{fileformat}")
+    file_path = Rails.root.join(filefolder, "#{fileroot}.#{fileformat}")
     File.read(file_path)
   end
